requirements.txt
Flask==2.0.3
Flask-RESTful==0.3.9
google-cloud-language
pytest